+++ Willowcat.CharacterGenerator.Model/Extension/SelectedOptionExtension.cs
@@ -1,0 +1,10 @@
+﻿namespace Willowcat.CharacterGenerator.Model.Extension
+{
+    public static class SelectedOptionExtension
+    {
+        public static string GetSelectionText(this IEnumerable<SelectedOption> options)
+        {
+            return string.Join(" - ", options.Select(option => option.Description));
+        }
+    }
+}
