+++ Willowcat.CharacterGenerator.UI/ViewModel/ChartViewModel.cs
@@ -7,7 +7,6 @@ using System.Threading.Tasks;
 using System.Windows.Documents;
 using System.Windows.Input;
 using Willowcat.CharacterGenerator.Core;
-using Willowcat.CharacterGenerator.Core.Models;
 using Willowcat.CharacterGenerator.Model;
 using Willowcat.CharacterGenerator.Model.Extension;
 using Willowcat.CharacterGenerator.UI.Commands;
